Changing Volatility to update more accurately
